Fix status bar color for Android

+++ shared/src/commonMain/kotlin/com/module/notelycompose/platform/expect/StatusBarManager.kt
@@ -1,6 +1,6 @@
 package com.module.notelycompose.platform.expect
 
 expect class StatusBarManager {
-    fun setStatusBarColor(color: Long)
+    fun setStatusBarColor(color: Int)
     fun restoreDefaultStatusBarColor()
 }
