+++ shared/src/iosMain/kotlin/com/module/notelycompose/platform/expect/StatusBarManager.ios.kt
@@ -2,7 +2,7 @@ package com.module.notelycompose.platform.expect
 
 actual class StatusBarManager {
 
-    actual fun setStatusBarColor(color: Long) {
+    actual fun setStatusBarColor(color: Int) {
 
     }
